requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
python-jose[all]
bcrypt